+++ datanymizer_engine/src/transformers/pipeline.rs
@@ -61,3 +61,35 @@ where
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        transformers::{CapitalizeTransformer, FirstNameTransformer, LastNameTransformer},
+        LocaleConfig, Transformers,
+    };
+
+    #[test]
+    fn defaults() {
+        let mut t = PipelineTransformer {
+            pipes: vec![
+                Transformers::FirstName(FirstNameTransformer::default()),
+                Transformers::LastName(LastNameTransformer {
+                    locale: Some(LocaleConfig::ZH_TW),
+                }),
+                Transformers::Capitalize(CapitalizeTransformer),
+            ],
+        };
+        t.set_defaults(&TransformerDefaults {
+            locale: LocaleConfig::RU,
+        });
+
+        assert!(
+            matches!(&t.pipes[0], Transformers::FirstName(t) if t.locale == Some(LocaleConfig::RU))
+        );
+        assert!(
+            matches!(&t.pipes[1], Transformers::LastName(t) if t.locale == Some(LocaleConfig::ZH_TW))
+        );
+    }
+}